requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
trl>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Configuration
pyyaml>=6.0

# Evaluation
scikit-learn>=1.3.0
numpy>=1.24.0

# Logging and monitoring
tensorboard>=2.14.0
wandb>=0.15.0  # Optional, for Weights & Biases logging

# Optimization (optional)
deepspeed>=0.10.0  # Optional, for DeepSpeed integration
flash-attn>=2.3.0  # Optional, for Flash Attention

# Utilities
tqdm>=4.66.0
scipy>=1.11.0
